(make_id_equals_comment): Cast sizeof expression (used as field width argument) to `int'.

+++ src/who.c
@@ -393,7 +393,7 @@ static char *
 make_id_equals_comment (STRUCT_UTMP const *utmp_ent)
 {
   char *comment = xmalloc (sizeof (_("id=")) + sizeof UT_ID (utmp_ent) + 1);
-  sprintf (comment, "%s%.*s", _("id="), sizeof UT_ID (utmp_ent),
+  sprintf (comment, "%s%.*s", _("id="), (int) sizeof UT_ID (utmp_ent),
 	   UT_ID (utmp_ent));
   return comment;
 }
